requirements.txt
PyQt6
requests
python-dotenv
signalrcore
Pillow
keyboard
websocket-client>=1.8.0,<2.0
pyaudio